rolekit-agent/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
openai==1.3.8
weasyprint==59.3
python-docx==0.8.11
celery==5.3.4
redis==5.0.1
alembic==1.13.0
jinja2==3.1.4

